orginizing done
